+++ src/main/java/com/pnu/momeet/domain/auth/dto/response/KakaoTokenResponse.java
@@ -1,0 +1,10 @@
+package com.pnu.momeet.domain.auth.dto.response;
+
+public record KakaoTokenResponse(
+        String access_token,
+        String token_type,
+        String refresh_token,
+        Integer expires_in,
+        String scope
+) {
+}
