+++ src/main/java/com/pnu/momeet/domain/auth/service/KakaoAuthService.java
@@ -2,8 +2,9 @@ package com.pnu.momeet.domain.auth.service;
 
 import com.pnu.momeet.common.exception.BannedAccountException;
 import com.pnu.momeet.common.security.util.JwtTokenProvider;
+import com.pnu.momeet.domain.auth.client.KakaoApiClient;
 import com.pnu.momeet.domain.auth.dto.KakaoUserInfo;
-import com.pnu.momeet.domain.auth.dto.request.KakaoTokenRequest;
+import com.pnu.momeet.domain.auth.dto.response.KakaoTokenResponse;
 import com.pnu.momeet.domain.auth.dto.response.TokenResponse;
 import com.pnu.momeet.domain.auth.entity.RefreshToken;
 import com.pnu.momeet.domain.auth.repository.RefreshTokenRepository;
@@ -14,14 +15,8 @@ import com.pnu.momeet.domain.member.service.MemberEntityService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.*;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.stereotype.Service;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -34,46 +29,22 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class KakaoAuthService {
 
-    @Value("${kakao.client-id}")
-    private String clientId;
-
-    @Value("${kakao.client-secret}")
-    private String clientSecret;
-
-    @Value("${kakao.redirect-uri}")
-    private String redirectUri;
-
-    @Value("${kakao.admin-key}")
-    private String adminKey;
-
-    @Value("${kakao.token-url}")
-    private String tokenUrl;
-
-    @Value("${kakao.user-info-url}")
-    private String userInfoUrl;
-
-    @Value("${kakao.unlink-url}")
-    private String unlinkUrl;
-
     private static final long IAT_BUFFER_SECONDS = 10;
 
-    private final RestTemplate restTemplate;
     private final MemberEntityService memberEntityService;
     private final JwtTokenProvider jwtTokenProvider;
     private final RefreshTokenRepository refreshTokenRepository;
+    private final KakaoApiClient kakaoApiClient;
 
     public String getKakaoAuthUrl() {
-        return UriComponentsBuilder.fromUriString("https://kauth.kakao.com/oauth/authorize")
-                .queryParam("response_type", "code")
-                .queryParam("client_id", clientId)
-                .queryParam("redirect_uri", redirectUri)
-                .build(true)
-                .toUriString();
+        return kakaoApiClient.getKakaoAuthUrl();
     }
 
     public KakaoUserInfo getKakaoUserInfo(String code) {
-        String token = getAccessTokenFromKakao(code);
-        return getKakaoUserInfoFromToken(token);
+        KakaoTokenResponse tokenResponse = kakaoApiClient.getAccessToken(code);
+        KakaoUserInfo userInfoResponse = kakaoApiClient.getUserInfo(tokenResponse.access_token());
+
+        return new KakaoUserInfo(userInfoResponse.kakaoId(), userInfoResponse.email());
     }
 
     public TokenResponse kakaoLogin(String code) {
@@ -85,75 +56,6 @@ public class KakaoAuthService {
 
         log.info("카카오 로그인 성공: {}", kakaoUserInfo.email());
         return tokenResponse;
-    }
-
-    private String getAccessTokenFromKakao(String code) {
-        KakaoTokenRequest request = new KakaoTokenRequest(
-                "authorization_code",
-                clientId,
-                clientSecret,
-                redirectUri,
-                code
-        );
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-        try {
-            String requestBody = convertToFormUrlEncoded(request);
-            HttpEntity<String> httpEntity = new HttpEntity<>(requestBody, headers);
-
-            ResponseEntity<Map> response = restTemplate.postForEntity(
-                    tokenUrl, httpEntity, Map.class
-            );
-
-            Map<String, Object> responseBody = response.getBody();
-            if (responseBody == null || !responseBody.containsKey("access_token")) {
-                throw new IllegalArgumentException("카카오에서 액세스 토큰을 받지 못했습니다.");
-            }
-
-            return (String) response.getBody().get("access_token");
-        } catch (Exception e) {
-            throw new IllegalArgumentException("카카오 인증 중 오류가 발생했습니다: " + e.getMessage());
-        }
-    }
-
-    private String convertToFormUrlEncoded(KakaoTokenRequest request) {
-        return String.format(
-                "grant_type=%s&client_id=%s&client_secret=%s&redirect_uri=%s&code=%s",
-                request.grant_type(),
-                request.client_id(),
-                request.client_secret(),
-                request.redirect_uri(),
-                request.code()
-        );
-    }
-
-    private KakaoUserInfo getKakaoUserInfoFromToken(String token) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setBearerAuth(token);
-
-        HttpEntity<String> request = new HttpEntity<>(headers);
-
-        try {
-            ResponseEntity<Map> response = restTemplate.exchange(userInfoUrl, HttpMethod.GET, request, Map.class);
-
-            Map<String, Object> userInfo = response.getBody();
-            if (userInfo == null || !userInfo.containsKey("id") || !userInfo.containsKey("kakao_account")) {
-                throw new IllegalArgumentException("카카오에서 사용자 정보를 받지 못했습니다.");
-            }
-
-            Map<String, Object> kakaoAccount = (Map<String, Object>) userInfo.get("kakao_account");
-            if (kakaoAccount == null || !kakaoAccount.containsKey("email")) {
-                throw new IllegalArgumentException("카카오 계정에서 이메일 정보를 받지 못했습니다.");
-            }
-
-            return new KakaoUserInfo(
-                    String.valueOf(userInfo.get("id")),
-                    (String) kakaoAccount.get("email")
-            );
-        } catch (Exception e) {
-            throw new IllegalArgumentException("카카오 사용자 정보 조회 중 오류가 발생했습니다: " + e.getMessage());
-        }
     }
 
     @Transactional
@@ -201,41 +103,6 @@ public class KakaoAuthService {
         return new TokenResponse(accessToken, refreshToken);
     }
 
-    private boolean callKakaoUnlinkApi(String kakaoId) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.set("Authorization", "KakaoAK " + adminKey);
-        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-
-        MultiValueMap<String, String> form = new LinkedMultiValueMap<>();
-        form.add("target_id_type", "user_id");
-        form.add("target_id", kakaoId);
-
-        HttpEntity<MultiValueMap<String, String>> request = new HttpEntity<>(form, headers);
-
-        try {
-            log.debug("카카오 연동 해제 API 요청: {}", unlinkUrl);
-            ResponseEntity<Map> response = restTemplate.postForEntity(unlinkUrl, request, Map.class);
-
-            if (!response.getStatusCode().is2xxSuccessful()) {
-                log.warn("카카오 연동 해제 API 실패: HTTP {} - {}", response.getStatusCode(), kakaoId);
-                return false;
-            }
-
-            Map<String, Object> responseBody = response.getBody();
-            boolean success = responseBody != null && responseBody.containsKey("id");
-
-            if (!success) {
-                log.warn("카카오 연동 해제 API 응답 오류: {}", responseBody);
-            }
-
-            log.info("카카오 연동 해제 API 성공: {}", kakaoId);
-            return success;
-        } catch (Exception e) {
-            log.error("카카오 연동 해제 실패: {}", e.getMessage());
-            return false;
-        }
-    }
-
     @Transactional
     public void withdrawKakaoMember(UUID memberId) {
         Member member = memberEntityService.getById(memberId);
@@ -246,7 +113,7 @@ public class KakaoAuthService {
         }
 
         log.info("카카오 연동 해제 API 호출 시작: {}", member.getProviderId());
-        boolean unlinkSuccess = callKakaoUnlinkApi(member.getProviderId());
+        boolean unlinkSuccess = kakaoApiClient.unlinkUser(member.getProviderId());
 
         if (!unlinkSuccess) {
             log.error("카카오 연동 해제 실패: {} ({})", member.getEmail(), member.getProviderId());
